Fix/minor api modify (#34)
* Modify ADC APIs


* Remove Newlib target


* Modify PWM api


* Modify GPIO & I2C api


* Modify SPI api


* Modify Timer api


* Modify UART api


* Modify Counter api


* Modify Platform api


* Modify OS api


* Modify OS api


* add include stdint.h


* Modify const semantics


* Modify fs API


* Modify i2s API


* Fix/modify obj to private (#35)

* Modify obj to private


* Modify obj to private


* Modify obj to private


---------


* Update api documentation (#36)

* Update api documentation

* Update digitalIn class explanation

---------

+++ Sources/CSwiftIO/include/swift_pwm.h
@@ -7,6 +7,9 @@
 #ifndef _SWIFT_PWM_H_
 #define _SWIFT_PWM_H_
 
+#include <stdint.h>
+#include <sys/types.h>
+
 /**
  * @brief Structure to receive pwm information
  *
@@ -14,8 +17,8 @@
  * @param min_frequency min pwm frequency
  */
 struct swift_pwm_info {
-	int max_frequency;
-	int min_frequency;
+	ssize_t max_frequency;
+	ssize_t min_frequency;
 };
 
 typedef struct swift_pwm_info swift_pwm_info_t;
@@ -48,7 +51,7 @@ int swifthal_pwm_close(void *pwm);
  * @retval 0 If successful.
  * @retval Negative errno code if failure.
  */
-int swifthal_pwm_set(void *pwm, int period, int pulse);
+int swifthal_pwm_set(void *pwm, ssize_t period, ssize_t pulse);
 
 /**
  * @brief Suspend pwm output
